Improve GWO Integrity
Add field representing current state of the algorithm. The idea is, that one can genereate an instance of the class with certain parameters. These parameters (as well as the ones describing current state of the algorithm) are stored and updated as private fields.
Divided initial Solve into smaller function in order to improve readibility. This provides better control over iterations and general state of the algorithm.